Changed wind variable type

+++ app/src/main/java/com/mikhailovskii/weatherandroid/data/entities/weather/Wind.kt
@@ -7,7 +7,7 @@ data class Wind(
 
     @SerializedName("speed")
     @Expose
-    var speed: Int? = null,
+    var speed: Double? = null,
 
     @SerializedName("deg")
     @Expose
